refactor: merge pointcloud loader module into core loader files
  * move PointcloudMapLoaderModule declaration/implementation into pointcloud_map_loader
  * remove obsolete pointcloud_map_loader_module header/source and CMake entry
  * update includes/call sites to use merged header
  * fix test include after transitive-header removal
