--- map/autoware_map_loader/src/pointcloud_map_loader/pointcloud_map_loader_module.hpp
@@ -1,53 +1,0 @@
-// Copyright 2022 The Autoware Contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef POINTCLOUD_MAP_LOADER__POINTCLOUD_MAP_LOADER_MODULE_HPP_
-#define POINTCLOUD_MAP_LOADER__POINTCLOUD_MAP_LOADER_MODULE_HPP_
-
-#include "pointcloud_map_loader.hpp"
-
-#include <rclcpp/rclcpp.hpp>
-
-#include <sensor_msgs/msg/point_cloud2.hpp>
-
-#include <boost/optional.hpp>
-
-#include <pcl/filters/voxel_grid.h>
-#include <pcl/io/pcd_io.h>
-#include <pcl_conversions/pcl_conversions.h>
-
-#include <string>
-#include <vector>
-
-namespace autoware::map_loader
-{
-class PointcloudMapLoaderModule
-{
-public:
-  PointcloudMapLoaderModule() = default;
-
-  // Backward-compatible constructor used by existing tests.
-  explicit PointcloudMapLoaderModule(
-    rclcpp::Node * node, const std::vector<std::string> & pcd_paths,
-    const std::string & publisher_name, const bool use_downsample);
-
-  static LoadPointcloudMapResult create_map_message(
-    const std::vector<std::string> & pcd_paths, boost::optional<float> leaf_size);
-
-private:
-  rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr pub_pointcloud_map_;
-};
-}  // namespace autoware::map_loader
-
-#endif  // POINTCLOUD_MAP_LOADER__POINTCLOUD_MAP_LOADER_MODULE_HPP_
